requirements.txt
streamlit==1.29.0
gspread==5.12.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
pandas==1.5.3
notion-client==2.2.1
st-gsheets-connection==0.0.3
streamlit-pills==0.3.0